requirements.txt
Django==5.1
django-markdownify==0.9.5
openai==1.61.0
python-dotenv==1.0.1
azure-core-tracing-opentelemetry==1.0.0b11
azure-search-documents==11.5.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-semantic-conventions==0.51b0
azure-monitor-opentelemetry==1.6.4
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-httpx==0.51b0
opentelemetry-instrumentation-openai-v2==2.1b0


